# server/src/config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "schemes": [
        "http"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get local articles",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "content": {
                    "application/json": {}
                },
                "parameters": []
            },
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "Update latest articles from Hacker News",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "content": {
                    "application/json": {}
                },
                "parameters": []
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Delete one local article",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "content": {
                    "application/json": {}
                },
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "query",
                        "description": "Specifies the article to delete",
                        "required": true,
                        "type": "number"
                    }
                ]
            }
        }
    }
}